Move private ip function into another class.
